# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.50
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/memory/l0:
    get:
      tags: [Memory]
      summary: List raw L0 messages
      description: Lists immutable raw messages with pagination, session, text, and recycle filters.
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: sessionId
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: includeDeleted
          in: query
          schema: { type: string, enum: [active, deleted, any], default: active }
      responses:
        "200":
          description: Paginated L0 messages
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/L0Turn" }
                  pagination: { $ref: "#/components/schemas/MemoryPagination" }
        "400": { description: Invalid query parameters }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Memory]
      summary: Import L0 messages or delete an L0 session
      description: |
        Without `sessionId`, imports 1–500 messages and returns 201. With
        `?sessionId=...`, the body must carry the same session id and the route
        soft- or permanently deletes that session, returning 200.
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: sessionId
          in: query
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [sessionId, items]
                  properties:
                    sessionId: { type: string, minLength: 1 }
                    items:
                      type: array
                      minItems: 1
                      maxItems: 500
                      items:
                        type: object
                        required: [idempotencyKey, role, content]
                        properties:
                          idempotencyKey: { type: string, minLength: 1, maxLength: 256 }
                          role: { type: string, enum: [user, assistant] }
                          content: { type: string, minLength: 1 }
                          timestamp: { type: string, format: date-time }
                          correlationId: { type: string }
                          provider: { type: string }
                          model: { type: string }
                - type: object
                  required: [sessionId]
                  properties:
                    sessionId: { type: string, minLength: 1 }
                    mode: { type: string, enum: [soft, permanent], default: soft }
      responses:
        "200": { description: L0 session deleted }
        "201": { description: L0 messages imported }
        "400": { description: Validation or storage error }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/memory/l1:
    get:
      tags: [Memory]
      summary: List or search L1 curated memories
      description: A non-empty `q` selects FTS5 search; otherwise this returns a listing.
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: sceneName
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/L1Type" }
        - name: q
          in: query
          schema: { type: string }
        - name: includeDeleted
          in: query
          schema: { type: string, enum: [active, deleted, any], default: active }
      responses:
        "200":
          description: Paginated L1 memories
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/L1Fact" }
                  pagination: { $ref: "#/components/schemas/MemoryPagination" }
        "400": { description: Invalid query parameters }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Memory]
      summary: Create an L1 curated memory
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, content, sceneName]
              properties:
                type: { $ref: "#/components/schemas/L1Type" }
                priority: { type: integer, minimum: 0, maximum: 100, default: 50 }
                content: { type: string, minLength: 1 }
                sceneName: { type: string, minLength: 1 }
                metadata: { type: object, additionalProperties: true, default: {} }
                sourceMessageIds:
                  type: array
                  maxItems: 500
                  items: { type: string }
                  default: []
      responses:
        "201": { description: L1 memory created }
        "400": { description: Validation or storage error }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/memory/l2:
    get:
      tags: [Memory]
      summary: List L2 derived scenes
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: sceneName
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: includeDeleted
          in: query
          schema: { type: string, enum: [active, deleted, any], default: active }
      responses:
        "200":
          description: Paginated L2 scenes
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/L2Entry" }
                  pagination: { $ref: "#/components/schemas/MemoryPagination" }
        "400": { description: Invalid query parameters }
    post:
      tags: [Memory]
      summary: Create an L2 scene
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sceneName, summary, heat, content]
              properties:
                sceneName: { type: string, minLength: 1 }
                groupKey: { type: [string, "null"] }
                summary: { type: string }
                heat: { type: number, minimum: 0, maximum: 1 }
                content: { type: string, minLength: 1 }
      responses:
        "201": { description: L2 scene created }
        "400": { description: Validation or storage error }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the OmniRoute dashboard
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie for protected management routes
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...
  schemas:
...
    MemoryPagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    L1Type:
      type: string
      enum: [persona, episodic, instruction, work_fact, work_task, work_method, work_artifact]
    L0Turn:
      type: object
      description: One immutable raw message in L0
      required:
        [
          id,
          ownerApiKeyId,
          sessionKey,
          role,
          content,
          timestamp,
          recordedAt,
          source,
          isInternal,
          truncated,
          idempotencyKey,
        ]
      properties:
        id: { type: string }
        ownerApiKeyId: { type: string }
        sessionKey: { type: string }
        sessionId: { type: [string, "null"] }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        timestamp: { type: string, format: date-time }
        recordedAt: { type: string, format: date-time }
        source: { type: string, enum: [user, assistant, imported] }
        correlationId: { type: [string, "null"] }
        comboExecutionKey: { type: [string, "null"] }
        isInternal: { type: boolean }
        provider: { type: [string, "null"] }
        model: { type: [string, "null"] }
        truncated: { type: boolean }
        idempotencyKey: { type: string }
        deletedAt: { type: [string, "null"], format: date-time }
    L1Fact:
      type: object
      description: One owner-curated L1 memory
      required:
        [
          id,
          ownerApiKeyId,
          type,
          priority,
          content,
          sceneName,
          metadata,
          sourceMessageIds,
          version,
          createdAt,
          updatedAt,
          lastModifiedBy,
          editedByUser,
        ]
      properties:
        id: { type: string }
        ownerApiKeyId: { type: string }
        type: { $ref: "#/components/schemas/L1Type" }
        priority: { type: integer, minimum: 0, maximum: 100 }
        content: { type: string }
        sceneName: { type: string }
        metadata: { type: object, additionalProperties: true }
        sourceMessageIds:
          type: array
          items: { type: string }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastModifiedBy: { type: string, enum: [user, pipeline] }
        editedByUser: { type: boolean }
        deletedAt: { type: [string, "null"], format: date-time }
    L2Entry:
      type: object
      description: One derived L2 scene
      required:
        [
          id,
          ownerApiKeyId,
          sceneName,
          summary,
          heat,
          content,
          version,
          createdAt,
          updatedAt,
          lastModifiedBy,
          editedByUser,
        ]
      properties:
        id: { type: string }
        ownerApiKeyId: { type: string }
        sceneName: { type: string }
        groupKey: { type: [string, "null"] }
        summary: { type: string }
        heat: { type: number, minimum: 0, maximum: 1 }
        content: { type: string }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastModifiedBy: { type: string, enum: [user, pipeline] }
        editedByUser: { type: boolean }
        deletedAt: { type: [string, "null"], format: date-time }
...